chore: rename swapper files without the 2
